Remove some redundant casts to (int) in some expressions.
git-svn-id: https://svn.code.sf.net/p/spacetraderjava/code@3 5295ef6b-c4f4-4cba-9fc2-ce1965aa78d7

+++ src/spacetrader/ShipSpec.java
@@ -355,7 +355,7 @@ public class ShipSpec extends STSerializableObject
 
 	public int ImageIndex()
 	{
-		return (_imageIndex == Consts.ShipImgUseDefault ? (int)Type().CastToInt() : _imageIndex);
+		return (_imageIndex == Consts.ShipImgUseDefault ? Type().CastToInt() : _imageIndex);
 	}
 
 	public void ImageIndex(int value)
